pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "screen"
version = "0.1.0"
description = "増配バリュー × ミネルヴィニ スクリーナー"
requires-python = ">=3.10"
dependencies = [
    "jquants-api-client>=2.0",
    "yfinance>=0.2",
    "pandas>=2.0",
    "numpy>=1.26",
    "requests>=2.31",
    "tenacity>=8.2",
    "diskcache>=5.6",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
screen = "screen.cli:run"

[tool.pytest.ini_options]
testpaths = ["tests"]
